--Moved template type-independent code to non-template base class.
This was motivated by "T.84: Use a non-template core implementation to provide an ABI-stable interface" in cpp core guidelines.  See https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rt-abi
